refactor dynamic tolerances
