+++ projects/hipdnn/test_sdk/include/hipdnn_test_sdk/utilities/DynamicTolerancesConv.hpp
@@ -1,0 +1,268 @@
+// Copyright © Advanced Micro Devices, Inc., or its affiliates.
+// SPDX-License-Identifier: MIT
+
+#pragma once
+
+#include <algorithm>
+#include <cmath>
+#include <limits>
+#include <stdexcept>
+#include <vector>
+
+#include <hipdnn_data_sdk/types.hpp>
+#include <hipdnn_test_sdk/utilities/DynamicTolerancesCommon.hpp>
+
+namespace hipdnn_test_sdk::utilities::conv
+{
+
+/**
+ * @brief Calculates the expected tolerance for Convolution Backward Weights (WrW) operations.
+ *
+ * This function estimates the maximum expected error due to floating-point accumulation during the
+ * computation of weight gradients. It considers the accumulation of products of inputs and output
+ * gradients over the batch and spatial dimensions.
+ *
+ * The tolerance is calculated by simulating the accumulation process using `ComputeType` precision
+ * and adding the precision loss from casting the final result to `OutputType`.
+ *
+ * Error Estimation Strategy:
+ * - High Precision (FP32, FP64): Uses a linear worst-case bound (Classical).
+ *   Error ≈ n * epsilon * maxProduct
+ * - Lower Precision (FP16, BF16): Uses a statistical/probabilistic bound.
+ *   Error ≈ k * sqrt(n) * epsilon * maxProduct
+ *
+ * It also accounts for precision loss if inputs are cast to a lower precision ComputeType.
+ *
+ * @tparam OutputType The data type of the output (weight gradients).
+ * @tparam InputType The data type of the input tensor.
+ * @tparam ComputeType The data type used for accumulation (default: float).
+ * @param inputMin The minimum value in the input tensor.
+ * @param inputMax The maximum value in the input tensor.
+ * @param dyMin The minimum value in the output gradient tensor.
+ * @param dyMax The maximum value in the output gradient tensor.
+ * @param dyDims The dimensions of the output gradient tensor (dy).
+ * @return The calculated tolerance value as float.
+ */
+template <typename OutputType, typename InputType, typename ComputeType = float>
+float calculateConvWrwTolerance(double inputMin,
+                                double inputMax,
+                                double dyMin,
+                                double dyMax,
+                                const std::vector<int64_t>& dyDims)
+{
+    hipdnn_test_sdk::utilities::validateComputeType<ComputeType>();
+
+    // dyDims: [N, K, Spatial...]
+    // Accumulation for weights (dw) happens over N and Spatial dimensions.
+    // dw[k, c, r, s] = sum_{n, h, w} dy[n, k, h, w] * x[n, c, h+r, w+s]
+
+    if(dyDims.empty() || dyDims.size() < 2)
+    {
+        throw std::invalid_argument("dyDims must have at least 2 dimensions (N, K).");
+    }
+
+    auto numberOfAccumulations = static_cast<uint64_t>(dyDims[0]); // Batch size N
+    for(size_t i = 2; i < dyDims.size(); ++i)
+    {
+        numberOfAccumulations *= static_cast<uint64_t>(dyDims[i]); // Spatial dimensions
+    }
+
+    const double maxAbsInput = std::max(std::abs(inputMin), std::abs(inputMax));
+    const double maxAbsDy = std::max(std::abs(dyMin), std::abs(dyMax));
+
+    // Worst case product magnitude
+    const double maxProduct = maxAbsInput * maxAbsDy;
+
+    // Bound on sum(|x_i * y_i|)
+    const double sumAbsProductBound = static_cast<double>(numberOfAccumulations) * maxProduct;
+
+    auto epsilon = static_cast<double>(std::numeric_limits<ComputeType>::epsilon());
+    const double gamma = hipdnn_test_sdk::utilities::computeGamma(numberOfAccumulations, epsilon);
+    hipdnn_test_sdk::utilities::validateGamma(gamma);
+
+    double accumulatedTolerance = gamma * sumAbsProductBound;
+
+    // Input casting error (2 distinct inputs: x and dy)
+    accumulatedTolerance
+        += hipdnn_test_sdk::utilities::computeInputCastingError<InputType, ComputeType>(
+            sumAbsProductBound);
+
+    // Output casting error
+    const double outputCastError
+        = hipdnn_test_sdk::utilities::computeOutputCastingError<OutputType, ComputeType>(
+            sumAbsProductBound);
+
+    const double totalTolerance = accumulatedTolerance + outputCastError;
+
+    hipdnn_test_sdk::utilities::validateToleranceRange<OutputType>(totalTolerance);
+
+    return static_cast<float>(totalTolerance);
+}
+
+/**
+ * @brief Calculates the expected tolerance for Convolution Backward Data (DGrad) operations.
+ *
+ * This function estimates the maximum expected error due to floating-point accumulation during the
+ * computation of input gradients. It considers the accumulation of products of output gradients
+ * and filter weights over the output channels and spatial filter dimensions.
+ *
+ * The tolerance is calculated by simulating the accumulation process using `ComputeType` precision
+ * and adding the precision loss from casting the final result to `OutputType`.
+ *
+ * Error Estimation Strategy:
+ * - High Precision (FP32, FP64): Uses a linear worst-case bound (Classical).
+ *   Error ≈ n * epsilon * maxProduct
+ * - Lower Precision (FP16, BF16): Uses a statistical/probabilistic bound.
+ *   Error ≈ k * sqrt(n) * epsilon * maxProduct
+ *
+ * It also accounts for precision loss if inputs are cast to a lower precision ComputeType.
+ *
+ * @tparam OutputType The data type of the output (input gradients dx).
+ * @tparam InputType The data type of the input tensors (dy and w).
+ * @tparam ComputeType The data type used for accumulation (default: float).
+ * @param dyMin The minimum value in the output gradient tensor (dy).
+ * @param dyMax The maximum value in the output gradient tensor (dy).
+ * @param wMin The minimum value in the filter weights tensor (w).
+ * @param wMax The maximum value in the filter weights tensor (w).
+ * @param wDims The dimensions of the filter weights tensor (w): [K, C, R, S] or [K, C, D, R, S].
+ * @return The calculated tolerance value as float.
+ */
+template <typename OutputType, typename InputType, typename ComputeType = float>
+float calculateConvDgradTolerance(
+    double dyMin, double dyMax, double wMin, double wMax, const std::vector<int64_t>& wDims)
+{
+    hipdnn_test_sdk::utilities::validateComputeType<ComputeType>();
+
+    // wDims: [K, C, Spatial...] for 2D: [K, C, R, S], for 3D: [K, C, D, R, S]
+    // Accumulation for input gradients (dx) happens over K (output channels) and Spatial filter dimensions.
+    // dx[n, c, h, w] = sum_{k, r, s} dy[n, k, p, q] * w[k, c, r, s]
+
+    if(wDims.empty() || wDims.size() < 4)
+    {
+        throw std::invalid_argument(
+            "wDims must have at least 4 dimensions for 2D convolution [K, C, R, S].");
+    }
+
+    // Number of accumulations = K * (product of spatial dimensions)
+    // For 2D: K * R * S
+    // For 3D: K * D * R * S
+    auto numberOfAccumulations = static_cast<uint64_t>(wDims[0]); // K (output channels)
+    for(size_t i = 2; i < wDims.size(); ++i)
+    {
+        numberOfAccumulations *= static_cast<uint64_t>(wDims[i]); // Spatial dimensions
+    }
+
+    const double maxAbsDy = std::max(std::abs(dyMin), std::abs(dyMax));
+    const double maxAbsW = std::max(std::abs(wMin), std::abs(wMax));
+
+    // Worst case product magnitude
+    const double maxProduct = maxAbsDy * maxAbsW;
+
+    // Bound on sum(|dy_i * w_i|)
+    const double sumAbsProductBound = static_cast<double>(numberOfAccumulations) * maxProduct;
+
+    auto epsilon = static_cast<double>(std::numeric_limits<ComputeType>::epsilon());
+    const double gamma = hipdnn_test_sdk::utilities::computeGamma(numberOfAccumulations, epsilon);
+    hipdnn_test_sdk::utilities::validateGamma(gamma);
+
+    double accumulatedTolerance = gamma * sumAbsProductBound;
+
+    // Input casting error (2 distinct inputs: dy and w)
+    accumulatedTolerance
+        += hipdnn_test_sdk::utilities::computeInputCastingError<InputType, ComputeType>(
+            sumAbsProductBound);
+
+    // Output casting error
+    const double outputCastError
+        = hipdnn_test_sdk::utilities::computeOutputCastingError<OutputType, ComputeType>(
+            sumAbsProductBound);
+
+    const double totalTolerance = accumulatedTolerance + outputCastError;
+
+    hipdnn_test_sdk::utilities::validateToleranceRange<OutputType>(totalTolerance);
+
+    return static_cast<float>(totalTolerance);
+}
+
+/**
+ * @brief Calculates the expected tolerance for Convolution Forward Propagation (fprop) operations.
+ *
+ * This function estimates the maximum expected error due to floating-point accumulation during the
+ * computation of forward convolution. It considers the accumulation of products of inputs and weights
+ * over the input channels and filter spatial dimensions.
+ *
+ * The tolerance is calculated by simulating the accumulation process using `ComputeType` precision
+ * and adding the precision loss from casting the final result to `OutputType`.
+ *
+ * Error Estimation Strategy:
+ * - High Precision (FP32, FP64): Uses a linear worst-case bound (Classical).
+ *   Error ≈ n * epsilon * maxProduct
+ * - Lower Precision (FP16, BF16): Uses a statistical/probabilistic bound.
+ *   Error ≈ k * sqrt(n) * epsilon * maxProduct
+ *
+ * It also accounts for precision loss if inputs are cast to a lower precision ComputeType.
+ *
+ * @tparam OutputType The data type of the output (forward convolution output).
+ * @tparam InputType The data type of the input tensor.
+ * @tparam ComputeType The data type used for accumulation (default: float).
+ * @param inputMin The minimum value in the input tensor.
+ * @param inputMax The maximum value in the input tensor.
+ * @param wMin The minimum value in the weight/filter tensor.
+ * @param wMax The maximum value in the weight/filter tensor.
+ * @param wDims The dimensions of the weight/filter tensor (w).
+ * @return The calculated tolerance value as float.
+ */
+template <typename OutputType, typename InputType, typename ComputeType = float>
+float calculateConvFpropTolerance(
+    double inputMin, double inputMax, double wMin, double wMax, const std::vector<int64_t>& wDims)
+{
+    hipdnn_test_sdk::utilities::validateComputeType<ComputeType>();
+
+    // wDims: [K, C, R, S]
+    // Accumulation for output (y) happens over C (input channels) and R, S (filter spatial dimensions).
+    // y[n, k, h, w] = sum_{c, r, s} x[n, c, h+r, w+s] * w[k, c, r, s]
+
+    if(wDims.empty() || wDims.size() < 2)
+    {
+        throw std::invalid_argument("wDims must have at least 2 dimensions (K, C).");
+    }
+
+    auto numberOfAccumulations = static_cast<uint64_t>(wDims[1]); // Input channels C
+    for(size_t i = 2; i < wDims.size(); ++i)
+    {
+        numberOfAccumulations *= static_cast<uint64_t>(wDims[i]); // Filter spatial dimensions R, S
+    }
+
+    const double maxAbsInput = std::max(std::abs(inputMin), std::abs(inputMax));
+    const double maxAbsW = std::max(std::abs(wMin), std::abs(wMax));
+
+    // Worst case product magnitude
+    const double maxProduct = maxAbsInput * maxAbsW;
+
+    // Bound on sum(|x_i * w_i|)
+    const double sumAbsProductBound = static_cast<double>(numberOfAccumulations) * maxProduct;
+
+    auto epsilon = static_cast<double>(std::numeric_limits<ComputeType>::epsilon());
+    const double gamma = hipdnn_test_sdk::utilities::computeGamma(numberOfAccumulations, epsilon);
+    hipdnn_test_sdk::utilities::validateGamma(gamma);
+
+    double accumulatedTolerance = gamma * sumAbsProductBound;
+
+    // Input casting error (2 distinct inputs: x and w)
+    accumulatedTolerance
+        += hipdnn_test_sdk::utilities::computeInputCastingError<InputType, ComputeType>(
+            sumAbsProductBound);
+
+    // Output casting error
+    const double outputCastError
+        = hipdnn_test_sdk::utilities::computeOutputCastingError<OutputType, ComputeType>(
+            sumAbsProductBound);
+
+    const double totalTolerance = accumulatedTolerance + outputCastError;
+
+    hipdnn_test_sdk::utilities::validateToleranceRange<OutputType>(totalTolerance);
+
+    return static_cast<float>(totalTolerance);
+}
+
+} // namespace hipdnn_test_sdk::utilities::conv
