+++ projects/hipdnn/test_sdk/tests/CMakeLists.txt
@@ -27,7 +27,9 @@ add_executable(
     utilities/TestCpuFpReferenceValidation.cpp
     utilities/TestCpuTypeUtilities.cpp
     utilities/TestCpuReferencePointwise.cpp
-    utilities/TestDynamicTolerances.cpp
+    utilities/TestDynamicTolerancesConv.cpp
+    utilities/TestDynamicTolerancesMatmul.cpp
+    utilities/TestDynamicTolerancesPointwise.cpp
     utilities/TestFlatbufferDatatypeMapping.cpp
     utilities/TestFlatbufferTensorAttributesUtils.cpp
     utilities/TestLoadGraphAndTensors.cpp
